EHTN_RFP_AGENT/requirements.txt
sentence-transformers
transformers
torch
faiss-cpu
pandas
openpyxl
